Move core_agraph functions to core/cagraph.c
